+++ utils/manifest-utils/Cargo.toml
@@ -32,6 +32,7 @@ md-5 = "0.10"
 log = "0.4"
 
 [dev-dependencies]
+tempfile = "3"
 anyhow = "1.0"
 serial_test = "0.9"
 shlex = "1.1"
